+++ lib/extension.ts
@@ -46,34 +46,108 @@ export function activate(context: ExtensionContext) {
 
     languages.setLanguageConfiguration("erlang", {
         onEnterRules: [
-            {
-                beforeText: /^.*(->|\s+(after|catch|if|of|receive))\s*$/,
-                action: { indentAction: IndentAction.Indent }
-            },
-            {
-                beforeText: /^\s*end$/,
-                action: { indentAction: IndentAction.Outdent }
-            },
-            {
-                beforeText: /^.*[^;,[({<]\s*$/,
-                action: { indentAction: IndentAction.Outdent }
-            },
-            {
-                beforeText: /^.*->.+;\s*$/,
-                action: { indentAction: IndentAction.None }
-            },
-            {
-                beforeText: /^.*(\.|;)\s*$/,
-                action: { indentAction: IndentAction.Outdent }
-            },
-            {
-                beforeText: /^%%% .*$/,
-                action: { indentAction: IndentAction.None, appendText: "%%% " }
-            },
-            {
-                beforeText: /^%%%.*$/,
-                action: { indentAction: IndentAction.None, appendText: "%%%" }
-            }
+            // Module comment: always continue comment
+            {
+                beforeText: /^%%% .*$/,
+                action: { indentAction: IndentAction.None, appendText: "%%% " }
+            },
+            {
+                beforeText: /^%%%.*$/,
+                action: { indentAction: IndentAction.None, appendText: "%%%" }
+            },
+            // Comment line with double %: continue comment if needed
+            {
+                beforeText: /^\s*%% .*$/,
+                afterText: /^.*\S.*$/,
+                action: { indentAction: IndentAction.None, appendText: "%% " }
+            },
+            {
+                beforeText: /^\s*%%.*$/,
+                afterText: /^.*\S.*$/,
+                action: { indentAction: IndentAction.None, appendText: "%%" }
+            },
+            // Comment line with single %: continue comment if needed
+            {
+                beforeText: /^\s*% .*$/,
+                afterText: /^.*\S.*$/,
+                action: { indentAction: IndentAction.None, appendText: "% " }
+            },
+            {
+                beforeText: /^\s*%.*$/,
+                afterText: /^.*\S.*$/,
+                action: { indentAction: IndentAction.None, appendText: "%" }
+            },
+            // Any other comment line: do nothing, ignore below rules
+            {
+                beforeText: /^\s*%.*$/,
+                afterText: /^\s*$/,
+                action: { indentAction: IndentAction.None }
+            },
+
+            // Empty line: do nothing, ignore below rules
+            {
+                beforeText: /^\s*$/,
+                action: { indentAction: IndentAction.None }
+            },
+
+            // Before guard sequence (before 'when')
+            {
+                beforeText: /.*/,
+                afterText: /^\s*when(\s.*)?$/,
+                action: { indentAction: IndentAction.None, appendText: "  " }
+            },
+            // After guard sequence (after 'when')
+            {
+                beforeText: /^\s*when(\s.*)?->\s*$/,
+                action: { indentAction: IndentAction.None, appendText: "  " }
+            },
+            {
+                beforeText: /^\s*when(\s.*)?(,|;)\s*$/,
+                action: { indentAction: IndentAction.None, appendText: "     " }
+            },
+
+            // Start of clause, right hand side of an assignment, after 'after', etc.
+            {
+                beforeText: /^.*(->|[^=]=|\s+(after|case|catch|if|of|receive|try))\s*$/,
+                action: { indentAction: IndentAction.Indent }
+            },
+
+            // Not closed bracket
+            {
+                beforeText: /^.*[(][^)]*$/,
+                action: { indentAction: IndentAction.Indent }
+            },
+            {
+                beforeText: /^.*[{][^}]*$/,
+                action: { indentAction: IndentAction.Indent }
+            },
+            {
+                beforeText: /^.*[[][^\]]*$/,
+                action: { indentAction: IndentAction.Indent }
+            },
+
+            // One liner clause but not the last
+            {
+                beforeText: /^.*->.+;\s*$/,
+                action: { indentAction: IndentAction.None }
+            },
+            // End of function or attribute (e.g. export list)
+            {
+                beforeText: /^.*\.\s*$/,
+                action: { indentAction: IndentAction.Outdent, removeText: 9000 }
+            },
+            // End of clause but not the last
+            // FIXME: After a guard (;) it falsely outdents
+            {
+                beforeText: /^.*;\s*$/,
+                action: { indentAction: IndentAction.Outdent }
+            },
+            // Last statement of a clause
+            // TODO: double outdent or outdent + removeText: <tabsize>
+            {
+                beforeText: /^.*[^;,[({<]\s*$/,
+                action: { indentAction: IndentAction.Outdent }
+            }
         ]
     });
 }